fscache: add debug logs via env `OSBUILD_DEBUG_LOGS=fscache`
This commit adds optional debug logs for the fscache via the
```
OSBUILD_DEBUG_LOGS=fscache
```
environment. We should probably include this into the main_cli
in the future and also integrate with `osbuild.Monitor`. This
can be done by writing an adapter that provides the interface
of a logging.Logger and bridges it to a osbuild.Monitor so
that it's part of the `osbuild.Monitor.log()` output. But that
is for a followup :) (KISS and all that).