tools/osbuild-depsolve-dnf5: repo info on package
When generating package sources and rpm stage metadata for a manifest
from a list of packages, we need to associate repository configuration
options to each package.

Add the information to the packages in osbuild-depsolve-dnf5 before
returning the packages so we don't have to do the association in the
user.  This is needed for reading repository configurations from a
directory so that the caller doesn't need to parse the configurations
itself.

(Copy of 97e105ed43cdff60a303d9457410a0da32b024ef to dnf5)