tools: merge osbuild-depsolve-dnf{,5} into a single file
There is a bit of identical code/logic in osbuild-depsolve-dnf{,5}
that is curently duplicated and while it has advantages (like that
there is a very clear dnf5 and dnf file) it also can lead to
drift over time. So I checked what it would take to share more
and here is my quick RFC. The basic idea is:
- use try/except ImportError to ensure we can work safely in envs
  wihtout dnf or dnf5
- use specialized SolverDNF{,5} solver classes (those are already
  there which is really nice) with a common base class (I have not
  checked if we actually need this, I wanted to get initial feedback
  first9
- that's really it :-)

To me this feels nice and worth considering as it allows us to
share more code.

There is more that could be done to make this commit nicer, the
exception handling could be unified more (it's a bit ugly right now),
there is probably potential to extract common helpers and maybe add
a few more tests along the way (the new tests already helped me a lot,
thanks for that!).

I'm happy to work more on this and bring it into a more mergable
state *if* people consider the idea valid.