workflow: add all missing tests by hand
This is an alternative version of PR#1735. It adds all the missing
tests manually. I do not like this because we will most likely
forget to them next time we add a new test cateogry (or a new file
in `test/run/*`). It seems also a bit wasteful because the "normal"
tests run in 4min so the setup cost for the worker/container checkout
etc seems just not worth it.