osbuild: add new top-level "from-container" json tag
This commit adds a new "from-container" json tag to the osbuild
manifest so that the given container is mounted as the buidlroot.

Note that this is just an experiment and needs more thinking how
to abstract/generalize this. It is meant as a faster way to do the
org.osbuild.deploy-container stage does.