stages: add "devices/mounts" as allowed bootc disks for tar
When moving to `bootc install to-filesystem` we will need support
for mounting the deployed disk and creating a tar from it for
WSL support. Note that we could also create a WSL from a bootc
container by just `install to-filesystem` and tar that (or use
`podman export`) but that would mean we wouldn't work with
our blueprints. So reusing the existing pipeline that writes
to a raw disk and taring seems a good way forward.