test: make appstream filelist empty
The appstream filelist is not needed for the test but it is 18MB
big. So instead of putting this into git just create an empty
filelist and update repomd.xml manually to match the new sha/size
etc. Note that dnf4 seems to only care about <checksum> and
<location> and we could omit the other fields. But its easy
enough so I added some more and if required we could add them
all.

Note also that we could simply remove the filelist from the repomd
and the "basic_module" test would still work, but there is a
different (part of a) test that assumes that every repo has a filelist
(`n_filelist_files == len(repo_servers_copy)`) and that would fail
if we omit the filelist. We could make this smarter too but just
shipping an empty filelist seems good enough.